requirements.txt
PySimpleGUI==3.9.1
python-chess==0.23.9
pystockfish==1.0.1
